+++ packages/venice-icons/CHANGELOG.md
@@ -3,6 +3,56 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/juntossomosmais/venice/compare/@juntossomosmais/venice-icons@0.4.0...@juntossomosmais/venice-icons@0.5.0) (2020-10-21)
+
+
+### Features
+
+* added two new icons ([65c56f4](https://github.com/juntossomosmais/venice/commit/65c56f4584373d6e161149a9441db821d22d3dab))
+
+
+
+
+
+# [0.4.0](https://github.com/juntossomosmais/venice/compare/@juntossomosmais/venice-icons@0.3.0...@juntossomosmais/venice-icons@0.4.0) (2020-10-06)
+
+
+### Features
+
+* added two icons from FA ([7a9e869](https://github.com/juntossomosmais/venice/commit/7a9e869601cd8a27efaea834e5c97346bdd311a3))
+
+
+
+
+
+# [0.3.0](https://github.com/juntossomosmais/venice/compare/@juntossomosmais/venice-icons@0.2.0...@juntossomosmais/venice-icons@0.3.0) (2020-09-30)
+
+
+### Bug Fixes
+
+* lint ([a2ef91a](https://github.com/juntossomosmais/venice/commit/a2ef91a5f367599fa57a9368765a2fb3f8c3a4e9))
+
+
+### Features
+
+* add padlock icon ([49f40b9](https://github.com/juntossomosmais/venice/commit/49f40b9c3f8eefa208d4e043bc889e551dde0439))
+
+
+
+
+
+# [0.2.0](https://github.com/juntossomosmais/venice/compare/@juntossomosmais/venice-icons@0.1.0...@juntossomosmais/venice-icons@0.2.0) (2020-09-29)
+
+
+### Features
+
+* added icons ([60819f8](https://github.com/juntossomosmais/venice/commit/60819f872aee2646335a9ca2ed57f073030119fc))
+* added svg icon to assets folder ([f65e920](https://github.com/juntossomosmais/venice/commit/f65e9201ed58fa4c7977f59fd46d841cb1ae7f4b))
+
+
+
+
+
 # 0.1.0 (2020-09-17)